Refactor OAuth into the base provider class
